Move user form to its own file from Users.tsx to UserForm.tsx
